Change type of longitude and latitude in Building

+++ src/main/java/com/erapulus/server/database/model/BuildingEntity.java
@@ -33,11 +33,11 @@ public class BuildingEntity implements Entity {
 
     @NotNull
     @Column("latitude")
-    private Long latitude;
+    private Double latitude;
 
     @NotNull
     @Column("longitude")
-    private Integer longitude;
+    private Double longitude;
 
     @NotNull
     @Column("university")
